add login and error handling

+++ lib/tasktracker_web/controllers/task_controller.ex
@@ -11,7 +11,12 @@ defmodule TasktrackerWeb.TaskController do
     render(conn, "index.json", tasks: tasks)
   end
 
-  def create(conn, %{"task" => task_params}) do
+  def create(conn, %{"task" => task_params, "token" => token_params}) do
+    {:ok, user_id} = Phoenix.Token.verify(conn, "auth token", token_params["token"])
+    if token_params["user_id"] != user_id do
+      raise "unauthourized"
+    end
+
     with {:ok, %Task{} = task} <- Tasks.create_task(task_params) do
       conn
       |> put_status(:created)
@@ -26,7 +31,12 @@ defmodule TasktrackerWeb.TaskController do
     render(conn, "show.json", task: task)
   end
 
-  def update(conn, %{"id" => id, "task" => task_params}) do
+  def update(conn, %{"id" => id, "task" => task_params, "token" => token_params}) do
+    {:ok, user_id} = Phoenix.Token.verify(conn, "auth token", token_params["token"])
+    if token_params["user_id"] != user_id do
+      raise "unauthourized"
+    end
+
     {id, _} = Integer.parse(id)
     task = Tasks.get_task!(id)
 
@@ -35,7 +45,13 @@ defmodule TasktrackerWeb.TaskController do
     end
   end
 
-  def delete(conn, %{"id" => id}) do
+  def delete(conn, %{"id" => id, "token" => token_params}) do
+    {:ok, user_id} = Phoenix.Token.verify(conn, "auth token", token_params["token"])
+    if token_params["user_id"] != user_id do
+      raise "unauthourized"
+    end
+
+    {id, _} = Integer.parse id
     task = Tasks.get_task!(id)
     with {:ok, %Task{}} <- Tasks.delete_task(task) do
       send_resp(conn, :no_content, "")
